2 pipelines for EasyOpenCV Sim
HSV checks a certain (predetermined) region of the screen for green hue and draws a box around it. It displays in the telemetry tab if it thinks there is a cup in the box.

SliderTest allows you to adjust sliders for HSV and blacks out all of the screen whose HSV values are not in the range of the sliders.